+++ src/main/java/com/project/cheerha/common/config/RedisConfig.java
@@ -20,8 +20,8 @@ public class RedisConfig {
         config.setCodec(new JsonJacksonCodec());
         config.useSingleServer()
 //                .setAddress("redis://redis:6379");  //도커로 실행할 경우
-//                .setAddress("redis://localhost:6379");    //로컬에서 실행할 경우
-               .setAddress("rediss://master.cheerha-redis.tlrpjs.apn2.cache.amazonaws.com:6379");   //aws
+                .setAddress("redis://localhost:6379");    //로컬에서 실행할 경우
+//                .setAddress("rediss://master.cheerha-redis.tlrpjs.apn2.cache.amazonaws.com:6379");   //aws
         return org.redisson.Redisson.create(config);
     }
