+++ src/main/java/com/project/cheerha/domain/user/entity/User.java
@@ -2,6 +2,8 @@ package com.project.cheerha.domain.user.entity;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
+import jakarta.persistence.EnumType;
+import jakarta.persistence.Enumerated;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
@@ -34,6 +36,7 @@ public class User {
     @Column(length = 255, nullable = false )
     private String password;
 
+    @Enumerated(EnumType.STRING)
     @Column(length = 5, nullable = false )
     private Role role;
